requirements.txt
requests
discord.py
audioop-lts
